python-memer/requirements.txt
praw==7.8.1
python-dotenv==1.1.1
requests==2.32.4
pillow==11.3.0
beautifulsoup4==4.13.4